+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Version 1.3.6
 -------------
+* Added additional index checks with tests.
 * Constant current constraint for CSC HVDC.
 * Net update reg Q participations.
 * In update set points, changed reg_by_gen to v_set_regulated.
@@ -25,6 +26,8 @@ Version 1.3.6
 * Added "rmpct" parameter to all regulating unit components: Gen, VSC, FACTS.
 * Added NET_update_reg_Q_participations and BUS_update_reg_Q_participations to update regulating Q participaptions based on RMPCT * (P/sum(P)).
 * Modified HEUR_PVPQ_SWITCHING_apply_step to update Q participations based on remaining devices that have reactive power capability.
+* Added shunt outage as a contingency structure.
+* Added bus outage as a contingency structure.
 
 Version 1.3.4
 -------------